refactoring, parse syntax for object
